+++ R/PLCI.mirt.R
@@ -21,7 +21,7 @@
 #'   this value will be adjusted accordingly
 #' @param lower logical; search for the lower CI?
 #' @param upper logical; search for the upper CI?
-#' @param NealeMiller logical; use the Neale and Miller 1997 approximation? Default is \code{FALSE}
+#' @param NealeMiller logical; use the Neale and Miller (1997) approximation? Default is \code{FALSE}
 #' @param verbose logical; include additional information in the console?
 #' @param ... additional arguments to pass to the estimation functions
 #'